refactor: GameUtil 과 GameValidator 클래스의 분리
SRP (단일 책임 원칙)를 고수하기 위해서 사용자의 입력을 검증 관련 기능들을 GameUtil에서 분리, GameValidator에 넣는 것으로 변경
GameValidator 테스트 클래스 추가 후 테스트 코드 이동
GameValidator 메소드 이름 직관적으로 변경 isCarNameInputValid, isLoopNumberInputValid
전진 조건에 해당하는 변수의 이름을 직관적으로 변경 MOVE_FORWARD -> MOVE_THRESHOLD

issue: https://github.com/JunHoPark93/java-racingcar/issues/9
