+++ src/main/java/com/junhopark/javaracingcar/domain/Car.java
@@ -19,7 +19,6 @@ public class Car {
         this.name = name;
     }
 
-    // 추가 기능 구현
     public void moveForward() {
         position++;
     }
